Move setting components in "settings" folder
